MNT: keep specific sim classes inside their respective modules
